tests/requirements.txt
pytest
python-lsp-server[all]
pyyaml
termcolor
tomlkit
typeguard
